Update -Penambahan Validasi Form -Fix bug id saat Insert/Update -Perubahan Tipe data ID dari Int menjadi Long

+++ app/src/main/java/com/example/geoform/helper/MappingHelper.kt
@@ -12,7 +12,7 @@ object MappingHelper {
         notesCursor?.apply {
             while (moveToNext()) {
                 val note = Note(
-                    id                    = getInt(getColumnIndexOrThrow(NoteColumns.ID)),
+                    id                    = getLong(getColumnIndexOrThrow(NoteColumns.ID)),
                     nomor_pole            = getString(getColumnIndexOrThrow(NoteColumns.NOMOR_POLE)),
                     tipe_tiang            = getString(getColumnIndexOrThrow(NoteColumns.TIPE_TIANG)),
                     equipment_tiang       = getString(getColumnIndexOrThrow(NoteColumns.EQUIPMENT_TIANG)),
